# Improve performance of LinearAlgebra.dot