--- a/src/interface.jl
+++ b/src/interface.jl
@@ -70,18 +70,38 @@ function promote_operation_fallback(
     return promote_operation(add_sub_op(op), x, y)
 end
 
+# ==============================================================================
+# The compiler in Julia 1.6 doesn't like `Vararg{Type,N}` so for performance
+# reasons we add an explicit method for the 2-argument case.
 function promote_operation_fallback(
     op::Union{AddSubMul,typeof(add_dot)},
-    T::Type,
+    ::Type{T},
+    ::Type{A},
+    ::Type{B},
+) where {T,A,B}
+    return promote_operation(
+        reduce_op(op),
+        T,
+        promote_operation(map_op(op), A, B),
+    )
+end
+
+function promote_operation_fallback(
+    op::Union{AddSubMul,typeof(add_dot)},
+    ::Type{T},
     args::Vararg{Type,N},
-) where {N}
+) where {T,N}
     return promote_operation(
         reduce_op(op),
         T,
         promote_operation(map_op(op), args...),
     )
 end
+# ==============================================================================
 
+# ==============================================================================
+# The compiler in Julia 1.6 doesn't like `Vararg{Type,N}` so for performance
+# reasons we add an explicit method for the 2-argument case.
 """
     promote_operation(op::Function, ArgsTypes::Type...)
 
@@ -91,6 +111,15 @@ the arguments `args` are `ArgsTypes`.
 function promote_operation(op::F, args::Vararg{Type,N}) where {F<:Function,N}
     return promote_operation_fallback(op, args...)
 end
+
+function promote_operation(
+    op::F,
+    ::Type{A},
+    ::Type{B},
+) where {F<:Function,A,B}
+    return promote_operation_fallback(op, A, B)
+end
+# ==============================================================================
 
 # Helpful error for common mistake
 function promote_operation(
@@ -215,6 +244,9 @@ type cannot be mutated to equal the result of the operation.
 """
 struct IsNotMutable <: MutableTrait end
 
+# ==============================================================================
+# The compiler in Julia 1.6 doesn't like `Vararg{Type,N}` so for performance
+# reasons we add an explicit method for the 2-argument case.
 """
     mutability(T::Type, ::typeof(op), args::Type...)::MutableTrait
 
@@ -228,6 +260,15 @@ function mutability(T::Type, op, args::Vararg{Type,N}) where {N}
         return IsNotMutable()
     end
 end
+
+function mutability(::Type{T}, op::F, ::Type{A}, ::Type{B}) where {T,F,A,B}
+    if mutability(T) isa IsMutable && promote_operation(op, A, B) == T
+        return IsMutable()
+    else
+        return IsNotMutable()
+    end
+end
+# ==============================================================================
 
 function mutability(x, op, args::Vararg{Any,N}) where {N}
     return mutability(typeof(x), op, typeof.(args)...)
@@ -632,10 +673,10 @@ promote_operation_fallback(::typeof(adjoint), a::Type) = a
 
 function promote_operation_fallback(
     ::typeof(LinearAlgebra.dot),
-    b::Type,
-    c::Type,
-)
-    return promote_operation(*, promote_operation(adjoint, b), c)
+    ::Type{A},
+    ::Type{B},
+) where {A,B}
+    return promote_operation(*, promote_operation(adjoint, A), B)
 end
 
 function buffer_for(::typeof(add_dot), a::Type, b::Type, c::Type)