Closes #132

As I suspected, the cause of #132 is some of the dispatch choices for when Julia specializes.

Before:
```julia
julia> @btime direct(100);
  476.903 ns (3 allocations: 2.67 KiB)

julia> @btime no_rewrite(100);
  19.890 μs (399 allocations: 262.59 KiB)

julia> @btime rewrite(100);
  30.507 μs (9 allocations: 5.14 KiB)

julia> @btime linear_algebra(100);
  96.442 μs (9 allocations: 5.12 KiB)
```
After (incl. https://github.com/jump-dev/MathOptInterface.jl/pull/1731):
```julia
julia> @btime direct(100);
  477.354 ns (3 allocations: 2.67 KiB)

julia> @btime no_rewrite(100);
  19.913 μs (399 allocations: 262.59 KiB)

julia> @btime rewrite(100);
  1.634 μs (9 allocations: 5.14 KiB)

julia> @btime manual(100);
  1.523 μs (9 allocations: 5.12 KiB)

julia> @btime linear_algebra(100);
  1.616 μs (9 allocations: 5.12 KiB)
```

This brings the linear algebra code back in line with manual code. But there's still some other things to investigate.

Code:
```Julia
using Revise

using BenchmarkTools
import LinearAlgebra
import MathOptInterface
import MutableArithmetics

const MOI = MathOptInterface
const MA = MutableArithmetics

function direct(n)
    d = 1.0:n
    x = MOI.VariableIndex.(1:n)
    return MOI.ScalarAffineFunction(MOI.ScalarAffineTerm.(d, x), 0.0)
end

function no_rewrite(n)
    d = 1.0:n
    x = MOI.VariableIndex.(1:n)
    return sum(d[i] * x[i] for i in 1:n)
end

function rewrite(n)
    d = 1.0:n
    x = MOI.VariableIndex.(1:n)
    return MA.@rewrite(sum(d[i] * x[i] for i in 1:n))
end

function linear_algebra(n)
    d = 1.0:n
    x = MOI.VariableIndex.(1:n)
    return LinearAlgebra.dot(d, x)
end

@btime direct(100);
@btime no_rewrite(100);
@btime rewrite(100);
@btime linear_algebra(100);
```