@odow 2022-02-08
O mate. This has to be my new all-time favorite performance bug. I have been smashing my head against this for hours.

As a sneak peek, here's what will happen once I update MOI:
```julia
julia> @btime direct(100);
  472.102 ns (3 allocations: 2.67 KiB)

julia> @btime no_rewrite(100);
  20.199 μs (399 allocations: 262.59 KiB)

julia> @btime rewrite(100);
  1.633 μs (9 allocations: 5.14 KiB)

julia> @btime manual(100);
  1.508 μs (9 allocations: 5.12 KiB)

julia> @btime linear_algebra(100);
  1.606 μs (9 allocations: 5.12 KiB)
```